Fixing up the database and numerous tweaks to following/unfollowing
I needed to fix the registration so that when a new user is created they also have empty arrays for followers and following. Previously i was running into issues where new test users would have no array so when i followed it'd update my following, but not the test users followers.