Made numerous changes to the backend
To clean things up and set ourselves up for data management. I creating a hosting array for users with references to the event documents they created. Now when user's delete their accounts, all information is removed from users to deleted_user's and all their hosted events are deleted and sent to deleted_events.

In the future if we add delete features for events, we can send the information there to retain data.

Also there is now a delete account functionality in settings.